backend/requirements.txt
fastapi==0.112.2
nbformat==5.10.4
pydantic==2.8.2
requests==2.32.3
uvicorn==0.30.6
python-dotenv==1.0.1
